+++ src/utils/arr/arr_get_set.c
@@ -1,12 +1,12 @@
 #include "arraydef.h"
 #include "memory.h"
 
-void	*arr_get(t_array *arr, unsigned int index)
+void	*arr_get(t_array *arr, unsigned long index)
 {
 	return (arr->elements + index * arr->bytes);
 }
 
-void	arr_set(t_array *arr, unsigned int index, void *element)
+void	arr_set(t_array *arr, unsigned long index, void *element)
 {
 	void	*ptr;
